requirements.txt
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.33.0
django-bootstrap-form==3.3
django-bootstrap3==9.0.0
django-ckeditor==5.3.0
django-js-asset==0.1.1
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
